Fetch License Files at Build Time from Their Official Websites
Official license files (Apache 2.0, EPL 2.0) are fetched during Maven build as text files
and used at runtime by the UI.
An additional "Proprietary" license is added with the text "proprietary". The text file is
stored in assets/licenses.
The generation of the copybara configuration file is changed so it is more robust (not using
magic numbers) but potentially less efficient. The reason is that this operation is only
executed once every month or two, and being fraction of seconds slower is fine regarding that it
is more robust.


Signed-off-by: Ghareeb Falazi <ghareeb.falazi@hotmail.com>