Pull allEntities deserialization from winery component into backend service
This more accurately reflects the responsibilities of the two classes, especially considering that winery-component is a UI component.
In future this should allow for clearer and easier modifications when dealing with the non-editable data cached by backendService.
Additionally this paves the way for DataType related requests from the PropertiesComponent without incurring an additional roundtrip.